+++ laktory/models/resources/databricks/dashboard.py
@@ -27,6 +27,10 @@ class Dashboard(BaseModel, PulumiResource, TerraformResource):
         Whether to embed credentials in the dashboard. Default is true.
     file_path:
         The path to the dashboard JSON file. Conflicts with serialized_dashboard.
+    name_prefix:
+        Prefix added to the dashboard display name
+    name_suffix:
+        Suffix added to the dashboard display name
     parent_path:
         The workspace path of the folder containing the dashboard. Includes leading slash and no trailing slash.
         If folder doesn't exist, it will be created.
@@ -65,21 +69,16 @@ class Dashboard(BaseModel, PulumiResource, TerraformResource):
     """
 
     access_controls: list[AccessControl] = []
-    # create_time: str = None
-    # dashboard_change_detected: bool = None
-    # dashboard_id: str = None
     dirpath: str = None
     display_name: str
     embed_credentials: bool = None
-    # etag: str = None
     file_path: str = None
-    # lifecycle_state: str = None
-    # md5: str = None
+    name_prefix: str = None
+    name_suffix: str = None
     parent_path: str = None
     path: str = None
     rootpath: str = None
     serialized_dashboard: str = None
-    # update_time: str = None
     warehouse_id: str
 
     @model_validator(mode="after")
@@ -105,9 +104,23 @@ class Dashboard(BaseModel, PulumiResource, TerraformResource):
 
         return self
 
+    @model_validator(mode="after")
+    def update_name(self) -> Any:
+        if self.name_prefix:
+            self.display_name = self.name_prefix + self.display_name
+            self.name_prefix = ""
+        if self.name_suffix:
+            self.display_name = self.display_name + self.name_suffix
+            self.name_suffix = ""
+        return self
+
     # ----------------------------------------------------------------------- #
     # Resource Properties                                                     #
     # ----------------------------------------------------------------------- #
+
+    @property
+    def resource_key(self) -> str:
+        return self.display_name
 
     @property
     def additional_core_resources(self) -> list[PulumiResource]:
@@ -135,7 +148,13 @@ class Dashboard(BaseModel, PulumiResource, TerraformResource):
 
     @property
     def pulumi_excludes(self) -> Union[list[str], dict[str, bool]]:
-        return ["access_controls", "dirpath", "rootpath"]
+        return [
+            "access_controls",
+            "dirpath",
+            "rootpath",
+            "name_prefix",
+            "name_suffix",
+        ]
 
     # ----------------------------------------------------------------------- #
     # Terraform Properties                                                    #